# backend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkillBridge API Ecosystem",
    "description": "Skill Intelligence and Career Readiness Platform API Specifications",
    "contact": {
      "name": "SkillBridge Development Team",
      "email": "support@skillbridge.internal"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://springdoc.org"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/skills/{skillId}/verify": {
      "post": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Verify student skill competency",
        "operationId": "verifySkill",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifySkillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentSkill"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/skills": {
      "get": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Get all available master skills",
        "operationId": "getSkills",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSkill"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills/{skillId}": {
      "get": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Get skill details by skillId",
        "operationId": "getSkillById",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSkill"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/skills/student/{skillId}": {
      "get": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Get student skill status for a specific skill",
        "operationId": "getStudentSkillById",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentSkill"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills/search": {
      "get": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Search skills by query and category",
        "operationId": "searchSkills",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSkill"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/skill/{skillId}": {
      "get": {
        "tags": [
          "Learning Recommendation Service"
        ],
        "summary": "Get courses for a specific skill",
        "operationId": "getCoursesBySkillId",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCourse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/skill/{skillId}": {
      "get": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Get evidence items for a specific skill",
        "operationId": "getEvidenceBySkillId",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListEvidence"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Evidence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "RESUME",
              "CERTIFICATE",
              "PROJECT",
              "ASSESSMENT",
              "INTERNSHIP",
              "PRACTICAL_TASK"
            ]
          },
          "fileUrl": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StudentSkill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "PARTIAL",
              "MISSING",
              "CLAIMED"
            ]
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "evidenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "evidenceListJson": {
            "type": "string"
          }
        }
      },
...
      "VerifySkillRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "score": {
            "type": "string"
          }
        }
      },
      "ApiResponseStudentSkill": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/StudentSkill"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListSkill": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skill"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ApiResponseSkill": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Skill"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "skillsCovered": {
            "type": "string"
          },
          "competenciesCovered": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "recommendationReason": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListEvidence": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListCourse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}